Refactor SAML2 plugin
The SAML2 plugin was really difficult to follow. It involved storing
values on the class object and flagging them for later use. There is a
possible threading issue there that I'm not sure is a real problem.

Refactor the plugin to be more functional so you can see what is
returned from each stage and how you use it.

Change-Id: I6783a8d58ec428e4a2813dde91cfa399dd1078f3